requirements.txt
pymupdf
sentence-transformers
chromadb
requests
python-dotenv
tqdm
streamlit
google-genai
